+++ first-web-app/src/main/java/ru/khrebtov/Lesson2/servlets/CatalogServlet.java
@@ -1,4 +1,4 @@
-package ru.khrebtov.Lesson3.servlets.errors;
+package ru.khrebtov.Lesson2.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,10 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(urlPatterns = "/not_found")
-public class ErrorNotFoundServlet extends HttpServlet {
+@WebServlet(name = "catalog_servlet", urlPatterns = "/catalog")
+public class CatalogServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().println("<h2>Страница не найдена, праверьте правильность запроса.</h2>");
+        resp.getWriter().println("<h1>Каталог товаров</h1>");
+        getServletContext().getNamedDispatcher("menu").include(req, resp);
     }
 }
