+++ first-web-app/src/main/java/ru/khrebtov/Lesson2/servlets/ProductServlet.java
@@ -1,4 +1,4 @@
-package ru.khrebtov.Lesson3.servlets;
+package ru.khrebtov.Lesson2.servlets;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,12 +7,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-
-@WebServlet(name = "main_servlet", urlPatterns = "/main")
-public class MainServlet extends HttpServlet {
+@WebServlet(name = "product_servlet", urlPatterns = "/product")
+public class ProductServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().println("<h1>Главная страница сайта.</h1>");
+        resp.getWriter().println("<h1>Продукт</h1>");
         getServletContext().getNamedDispatcher("menu").include(req, resp);
     }
 }
