+++ first-web-app/src/main/java/ru/khrebtov/Lesson2/servlets/errors/ErrorNotFoundServlet.java
@@ -1,4 +1,4 @@
-package ru.khrebtov.Lesson3.servlets;
+package ru.khrebtov.Lesson2.servlets.errors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "product_servlet", urlPatterns = "/product")
-public class ProductServlet extends HttpServlet {
+@WebServlet(urlPatterns = "/not_found")
+public class ErrorNotFoundServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().println("<h1>Продукт</h1>");
-        getServletContext().getNamedDispatcher("menu").include(req, resp);
+        resp.getWriter().println("<h2>Страница не найдена, праверьте правильность запроса.</h2>");
     }
 }
